Added helper methods to Level Control interface.
Created methods that receive new level in the range of [0,100] and
convert the value to the interface's interval [0,255].

+++ tests/interfaces/test_level_control.cpp
@@ -68,7 +68,7 @@ TEST (LevelControlClient, Level)
 {
    mock ("Interface").expectOneCall ("send");
 
-   client.level (addr, 0x42);
+   client.level (addr, (uint8_t) 0x42);
 
    mock ("Interface").checkExpectations ();
 
@@ -134,7 +134,7 @@ TEST (LevelControlServer, Level)
 {
    CHECK_EQUAL (0, server.level ());
    mock ("Interface").expectOneCall ("notify");
-   server.level (42);
+   server.level ((uint8_t) 42);
    mock ("Interface").checkExpectations ();
    CHECK_EQUAL (42, server.level ());
 }
